Add files via upload
A Loop is a control structure that allows to execute a block of code repeatedly for a specified number of iterations. There are two main types of loops in Python (For & While  Loop).
For loop used to iterate over a sequence and execute a block of code for each items.
While loop  used to iterate over a sequence and execute a block of code as a specified condition is true.